Add simple JSGI style response handler
This allows any of the routes to return a standard JSGI style response
and have it parsed into the correct response.  Makes node-router a lot
more versatile in its automatic response handling from my point-of-view.

There's still a lot of room for improvement, I'm sure.  This is a first
pass at it late at night after the idea struck me.